trim lower upper insert tests (need fix)

+++ s21_string_lib/s21_to_lower.c
@@ -6,7 +6,7 @@ void *s21_to_lower(const char *str) {
   if (str == S21_NULL) {
     result = S21_NULL;
   } else {
-    result = calloc(s21_strlen(str), sizeof(char));
+    result = calloc((size_t)s21_strlen(str), sizeof(char));
     int flag = 1;
     for (s21_size_t i = 0; i < s21_strlen(str) && flag == 1; i++) {
       if (str[i] >= 'A' && str[i] <= 'Z') {
